+++ ts/block/parent_block.ts
@@ -1,4 +1,4 @@
-import { Popup } from "../popup.js";
+import { Popup } from "../popup/popup.js";
 import { BasicBlock } from "./basic_block.js";
 import { EmptyBlock } from "./concrete_block/empty_block.js";
 import { Roots } from "./roots.js";
@@ -32,7 +32,7 @@ export abstract class ParentBlock extends BasicBlock {
     }
 
     protected popup_event_kill_self() {
-        Popup.remove_popup();
+        Popup.remove_all_popup();
         const children_cnt = this.get_children().length;
         this.disconnect_all_children();
         for (let i = 0; i < children_cnt; ++i) {
